--- src/models/user.ts
@@ -1,80 +1,0 @@
-/*************************************************************************
- * Copyright 2020 Gravwell, Inc. All rights reserved.
- * Contact: <legal@gravwell.io>
- *
- * This software may be modified and distributed under the terms of the
- * MIT license. See the LICENSE file for details.
- **************************************************************************/
-
-import { isBoolean, isDate, isString, isNull } from 'lodash';
-import { NumericID, isNumericID } from '../value-objects';
-
-export interface User {
-	id: NumericID;
-	groupIDs: Array<NumericID>;
-
-	username: string;
-	name: string;
-	email: string;
-	role: 'admin' | 'analyst';
-	locked: boolean;
-	lastActivityDate: Date | null;
-}
-
-export type UserRole = User['role'];
-
-export interface RawUser {
-	UID: number;
-	User: string;
-	Name: string;
-	Email: string;
-	Admin: boolean;
-	Locked: boolean;
-	TS: string;
-
-	Synced: boolean;
-	Groups: Array<{
-		GID: number;
-		Name: string;
-		Desc: string;
-		Synced: boolean;
-	}>;
-}
-
-export const toUser = (raw: RawUser): User => ({
-	id: raw.UID.toString(),
-	groupIDs: raw.Groups.map(g => g.GID.toString()),
-	username: raw.User,
-	name: raw.Name,
-	email: raw.Email,
-	role: raw.Admin ? 'admin' : 'analyst',
-	locked: raw.Locked,
-	lastActivityDate: ((): Date | null => {
-		// !WARNING: If we set the TS to January 01, 0001, it means that there's no last activity date
-		// I'm using 1970 to have a margin for errors
-		const nullDate = new Date(10000000000);
-		const date = new Date(raw.TS);
-		return date.getTime() <= nullDate.getTime() ? null : date;
-	})(),
-});
-
-export const isValidUserRole = (value: any): value is UserRole =>
-	(<Array<UserRole>>['admin', 'analyst']).includes(value);
-
-export const isValidUser = (value: any): value is User => {
-	try {
-		const u = <User>value;
-		return (
-			isNumericID(u.id) &&
-			u.groupIDs.every(isNumericID) &&
-			isString(u.username) &&
-			isString(u.name) &&
-			isString(u.email) &&
-			isValidUserRole(u.role) &&
-			isBoolean(u.locked) &&
-			(isDate(u.lastActivityDate) || isNull(u.lastActivityDate))
-		);
-	} catch {
-		return false;
-	}
-};
